pyproject.toml
[project]
name = "mister-x-web-ng"
version = "0.1.0"
description = "WebApp to use for running a MisterX game"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crispy-bootstrap5>=2024.10",
    "django>=5.1.7",
    "django-crispy-forms>=2.3",
    "django-debug-toolbar>=5.0.1",
    "django-filter>=25.1",
    "django-guardian>=2.4.0",
    "django-tables2>=2.7.5",
    "psycopg2-binary>=2.9.10",
    "social-auth-app-django>=5.4.3",
    "uwsgi>=2.0.28",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.1.3",
    "djhtml>=3.0.7",
    "djlint>=1.36.4",
    "mypy>=1.15.0",
    "ruff>=0.9.10",
]

[[tool.mypy.overrides]]
module = [
    "crispy_forms.*",
    "debug_toolbar.*",
    "django_tables2.*",
    "guardian.*",
    "django_filters.*",
]
ignore_missing_imports = true

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.ruff]
lint.extend-select = ["I", "UP", "FURB"]
line-length = 129
exclude = ["migrations"]

[tool.django-stubs]
django_settings_module = "misterx_root.settings"

[tool.djlint]
blank_line_before_tag="extends,block"
blank_line_after_tag="load,extends,endblock"
format_attribute_template_tags=true
format_css=true
format_js=true
# Ignore meta description and tags
ignore="H030,H031"
max_blank_lines=1
profile="django"
max_attribute_length=80


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/misterx", "src/misterx_root", "src/templates", "src/utilities"]
